Fix read_locktime_from_contract()
Previously this function would only work if the locktime was a 16 bit
number. If the value would fit in one byte the function would return
as wrong result.

Also changed the type of locktime to be u16, which is more correct as
the OP_CSV bip talks about it only being 16 bits.

+++ src/messages.rs
@@ -38,7 +38,7 @@ pub struct SenderContractTxNoncesInfo {
 pub struct SignSendersContractTx {
     pub txes_info: Vec<SenderContractTxNoncesInfo>,
     pub hashvalue: [u8; 20],
-    pub locktime: i64,
+    pub locktime: u16,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
@@ -61,7 +61,7 @@ pub struct NextCoinSwapTxInfo {
 pub struct ProofOfFunding {
     pub confirmed_funding_txes: Vec<ConfirmedCoinSwapTxInfo>,
     pub next_coinswap_info: Vec<NextCoinSwapTxInfo>,
-    pub next_locktime: i64,
+    pub next_locktime: u16,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
